pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "RAG pipeline over vLLM documentation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bm25>=0.3.8",
    "fire",
    "numpy>=2.2.6",
    "pydantic",
]

[project.scripts]
student = "student.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/student"]
